+++ changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Platforms content and component
 - Portfolio configuration, content and functionality
 - Experience configuration, content and functionality
+- Education configuration, content and functionality
 - 
 ## [0.0.0] - 2019-01-14